cleanup gui tests
