+++ tests/SmartMT/__init__.py
@@ -1,10 +1,13 @@
 from __future__ import print_function
 
 import os
+import random
 import sys
 
 import matplotlib
 import sip
+from qtpy import QtCore
+from qtpy.QtTest import QTest
 from qtpy.QtWidgets import QApplication
 
 from mtpy.utils.mtpylog import MtPyLog
@@ -25,15 +28,13 @@ else:
 MtPyLog().get_mtpy_logger(__name__).info("Testing using matplotlib backend {}".format(matplotlib.rcParams['backend']))
 
 
-def reset_matplotlib():
-    # save some important params
-    interactive = matplotlib.rcParams['interactive']
-    backend = matplotlib.rcParams['backend']
-    # reset
-    matplotlib.rcdefaults()  # reset the rcparams to default
-    # recover
-    matplotlib.rcParams['backend'] = backend
-    matplotlib.rcParams['interactive'] = interactive
-    logger = MtPyLog().get_mtpy_logger(__name__)
-    logger.info("Testing using matplotlib backend {}".format(matplotlib.rcParams['backend']))
-
+def _click_area(qobj, pos=None, offset=None):
+    geom = qobj.geometry()
+    if pos is None:
+        x = int(geom.width() * random.uniform(0.2, 0.8))  # avid to click on the edge of widgets
+        y = int(geom.height() * random.uniform(0.2, 0.8))
+        pos = QtCore.QPoint(x, y)
+    if offset is not None:
+        pos += offset
+    QTest.mouseClick(qobj, QtCore.Qt.LeftButton, pos=pos)
+    print(pos.x(), pos.y())
