--- tests/SmartMT/gui/__init__.py
@@ -1,16 +1,0 @@
-import random
-
-from qtpy import QtCore
-from qtpy.QtTest import QTest
-
-
-def _click_area(qobj, pos=None, offset=None):
-    geom = qobj.geometry()
-    if pos is None:
-        x = int(geom.width() * random.uniform(0.2, 0.8))  # avid to click on the edge of widgets
-        y = int(geom.height() * random.uniform(0.2, 0.8))
-        pos = QtCore.QPoint(x, y)
-    if offset is not None:
-        pos += offset
-    QTest.mouseClick(qobj, QtCore.Qt.LeftButton, pos=pos)
-    print(pos.x(), pos.y())
